Expose parsing of input data
Allows the external context (for example, a controller) to integrate
deserialize and serialize actions.

When serializing, we want to return data that was changed, so that a
consumer can just replace their version with the updated server version
and *see their changes*. For this we need to reflect on what was
changed, (or equivalently what's going to be changed), and incorporate
this into the serialize_context.

Since the data is currently internal to `deserialize_from_view` we need
to bring it out into the caller. Unforunately this means a bunch of
reimplementation effort at each callsite.