Return changed associations on update
Associations may be optional, which will prevent them from being
rendered by default. When serializing, we want to return data that was
changed, so that a consumer can just replace their version with the
updated server version and *see their changes*. We infer and specify the
`include:` for the serialize context by inspecting the update data.