Record `changed?` when asserting a mutable value on creation
Even if it's the model's default value. This allows callbacks and access control
to differentiate between an explicitly asserted value and a fallback to the
model default.

Also refactors the dummy model in `RecordTest`: previously the "model default
values" were anything but that, and couldn't differentiate between a value that
was asserted and that was not.