Use both new and old schema versions when calculating migrated etag
To identify a migrated schema version for caching purposes, we need to use both
the current and target schema versions. Otherwise if we were to make future
migrations that would affect the result when migrating to older views (e.g. by
discarding and reconstructing data), the cache would not be invalidated and
cached results would be different from computed ones.