Allow migrations to create a references section
The 'references' section only exists in a serialization if it contains
some data. Migrations are always provided a valid references hash to
mutate. Mutations of the hash were discarded even when the migration
caused the hash to be non-empty.

Additionally the references hash might be emptied out by a migration,
but would have still been returned.