[API change] Vacate implicit cast #to_hash method on ViewModel
ViewModel defined the to_hash instance method to serialize the view to a hash.
This really isn't something we wanted to do, because to_hash is the implicit
type coercion method, and will be called by Ruby when we don't expect it.

Rename the method to #serialize_to_hash to match the corresponding class method.