During rendering, allow yielded block to contribute references
Reference serialization in prerender_viewmodel/prerender_json_view are is now
deferred until the optional metadata block. The block is now passed a serialize
context, and it may contribute references to this block that will be included.

SerializeContext gains a new `add_preserialized_reference` to permit recording a
pre-serialized reference literal.