Add way for views wrapping AR views to participate in preloading
A non-AR viewmodel that wraps a collection of AR viewmodels is a pattern we
occasionally use. In cases like this, it's very easy to accidentally miss
preloading, because `ViewModel.preload_for_serialization` will do nothing.

Add a `preloadable_dependencies` method to `ViewModel`, which allows a view to
specify one or more AR views that can be `eager_includes`-preloaded. For an AR
view, that would be `self`, and for a wrapper, it would be the wrapped AR views.